reorganize components + light mode styles
