+++ frontend/src/pages/WorkspaceSettings/ChatSettings/ChatPromptSettings/ChatPromptHistory/index.jsx
@@ -1,9 +1,9 @@
-import { useEffect, useState, forwardRef, useRef } from "react";
+import { useEffect, useState, forwardRef } from "react";
 import PromptHistory from "@/models/promptHistory";
-import { DotsThreeVertical, X } from "@phosphor-icons/react";
-import moment from "moment";
+import { X } from "@phosphor-icons/react";
+import PromptHistoryItem from "./PromptHistoryItem";
 
-const ChatPromptHistory = forwardRef(function ChatPromptHistory(
+export default forwardRef(function ChatPromptHistory(
   { show, workspaceSlug, onRestore, onClose },
   ref
 ) {
@@ -42,7 +42,7 @@ const ChatPromptHistory = forwardRef(function ChatPromptHistory(
   return (
     <div
       ref={ref}
-      className={`fixed right-3 top-3 bottom-3 w-[375px] bg-theme-action-menu-bg rounded-xl py-4 px-4 z-[9999] overflow-y-auto ${
+      className={`fixed right-3 top-3 bottom-3 w-[375px] bg-theme-action-menu-bg light:bg-theme-home-update-card-bg rounded-xl py-4 px-4 z-[9999] overflow-y-auto ${
         show
           ? "translate-x-0 opacity-100 visible"
           : "translate-x-full opacity-0 invisible"
@@ -54,7 +54,7 @@ const ChatPromptHistory = forwardRef(function ChatPromptHistory(
         </div>
         <div className="flex items-center gap-2">
           <div
-            className="text-sm font-medium text-white cursor-pointer hover:text-gray-300"
+            className="text-sm font-medium text-white cursor-pointer hover:text-primary-button"
             onClick={handleClearAll}
           >
             Clear All
@@ -62,7 +62,7 @@ const ChatPromptHistory = forwardRef(function ChatPromptHistory(
           <X
             size={16}
             weight="bold"
-            className="text-white cursor-pointer hover:text-gray-300"
+            className="text-white cursor-pointer hover:text-primary-button"
             onClick={onClose}
           />
         </div>
@@ -87,101 +87,3 @@ const ChatPromptHistory = forwardRef(function ChatPromptHistory(
     </div>
   );
 });
-
-function PromptHistoryItem({
-  id,
-  prompt,
-  modifiedAt,
-  user,
-  onRestore,
-  setHistory,
-}) {
-  const [showMenu, setShowMenu] = useState(false);
-  const menuRef = useRef(null);
-  const menuButtonRef = useRef(null);
-
-  const deleteHistory = async (id) => {
-    if (
-      confirm(
-        "Are you sure you want to delete this history item? This action cannot be undone."
-      )
-    ) {
-      const { success } = await PromptHistory.delete(id);
-      if (success) {
-        setHistory((prevHistory) =>
-          prevHistory.filter((item) => item.id !== id)
-        );
-      }
-    }
-  };
-
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (
-        showMenu &&
-        !menuRef.current.contains(event.target) &&
-        !menuButtonRef.current.contains(event.target)
-      ) {
-        setShowMenu(false);
-      }
-    };
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, [showMenu]);
-
-  return (
-    <div className="text-white">
-      <div className="flex items-center justify-between">
-        <div className="text-xs">
-          {user && (
-            <>
-              <span className="text-sky-400">{user.username}</span>{" "}
-              <span className="mx-1 text-white">•</span>
-            </>
-          )}
-          <span className="text-[#B6B7B7]">{moment(modifiedAt).fromNow()}</span>
-        </div>
-        <div className="flex items-center gap-2">
-          <div
-            className="text-xs cursor-pointer hover:text-gray-300"
-            onClick={onRestore}
-          >
-            Restore
-          </div>
-          <div className="relative">
-            <div
-              ref={menuButtonRef}
-              className="text-white cursor-pointer hover:text-gray-300"
-              onClick={() => setShowMenu(!showMenu)}
-            >
-              <DotsThreeVertical size={16} weight="bold" />
-            </div>
-            {showMenu && (
-              <div
-                ref={menuRef}
-                className="absolute right-0 top-6 bg-black rounded-lg z-50"
-              >
-                <div
-                  className="px-[10px] py-[6px] text-xs text-white hover:bg-theme-hover cursor-pointer"
-                  onClick={() => {
-                    setShowMenu(false);
-                    deleteHistory(id);
-                  }}
-                >
-                  Delete
-                </div>
-              </div>
-            )}
-          </div>
-        </div>
-      </div>
-      <div className="flex items-center mt-1">
-        <div className="text-white text-sm font-medium">{prompt}</div>
-      </div>
-    </div>
-  );
-}
-
-export default ChatPromptHistory;
